requirements.txt
pyrofork
motor
TgCrypto
Dnspython
aiohttp
pytz
pillow